+++ doctor-plugin/src/main/java/com/osacky/doctor/DoctorPlugin.kt
@@ -48,14 +48,15 @@ class DoctorPlugin : Plugin<Project> {
         val javaHomeCheck = JavaHomeCheck(extension, pillBoxPrinter)
         val garbagePrinter = GarbagePrinter(clock, DirtyBeanCollector(), extension)
         val buildOperations = getOperationEvents(target, extension)
-        val javaAnnotationTime = JavaAnnotationTime(buildOperations, extension, target.buildscript.configurations)
+        val javaAnnotationTime = JavaAnnotationTime(buildOperations, extension)
         val downloadSpeedMeasurer = DownloadSpeedMeasurer(buildOperations, extension, intervalMeasurer)
         val buildCacheConnectionMeasurer = BuildCacheConnectionMeasurer(buildOperations, extension, intervalMeasurer)
         val buildCacheKey = RemoteCacheEstimation((buildOperations as BuildOperations), target, clock)
         val slowerFromCacheCollector = buildOperations.slowerFromCacheCollector()
         val jetifierWarning = JetifierWarning(extension, target)
         val javaElevenGC = JavaGCFlagChecker(pillBoxPrinter, extension)
-        val list = listOf(daemonChecker, javaHomeCheck, garbagePrinter, javaAnnotationTime, downloadSpeedMeasurer, buildCacheConnectionMeasurer, buildCacheKey, slowerFromCacheCollector, jetifierWarning, javaElevenGC)
+        val kotlinCompileDaemonFallbackDetector = KotlinCompileDaemonFallbackDetector(target, extension)
+        val list = listOf(daemonChecker, javaHomeCheck, garbagePrinter, javaAnnotationTime, downloadSpeedMeasurer, buildCacheConnectionMeasurer, buildCacheKey, slowerFromCacheCollector, jetifierWarning, javaElevenGC, kotlinCompileDaemonFallbackDetector)
 
         garbagePrinter.onStart()
         javaAnnotationTime.onStart()
@@ -67,6 +68,7 @@ class DoctorPlugin : Plugin<Project> {
             daemonChecker.onStart()
             javaHomeCheck.onStart()
             javaElevenGC.onStart()
+            kotlinCompileDaemonFallbackDetector.onStart()
         }
 
         val buildScanApi = ScanApi(target)
@@ -155,7 +157,7 @@ class DoctorPlugin : Plugin<Project> {
             pillBoxPrinter.writePrescription(thingsToPrint)
         }
 
-        if (target.gradle.shouldUseCoCaClasses()) {
+        if (shouldUseCoCaClasses()) {
             val closeService =
                 target.gradle.sharedServices.registerIfAbsent("close-service", BuildFinishService::class.java) { }.get()
             closeService.closeMeWhenFinished {
@@ -198,8 +200,8 @@ class DoctorPlugin : Plugin<Project> {
     }
 
     private fun ensureMinimumSupportedGradleVersion() {
-        if (GradleVersion.current() < GradleVersion.version("5.2")) {
-            throw GradleException("Must be using Gradle Version 5.2 in order to use DoctorPlugin. Current Gradle Version is ${GradleVersion.current()}")
+        if (GradleVersion.current() < GradleVersion.version("6.1.1")) {
+            throw GradleException("Must be using Gradle Version 6.1.1 in order to use DoctorPlugin. Current Gradle Version is ${GradleVersion.current()}")
         }
     }
 
@@ -211,7 +213,7 @@ class DoctorPlugin : Plugin<Project> {
     }
 
     private fun getOperationEvents(target: Project, extension: DoctorExtension): OperationEvents {
-        return if (target.gradle.shouldUseCoCaClasses()) {
+        return if (shouldUseCoCaClasses()) {
             val listenerService = target.gradle.sharedServices.registerIfAbsent("listener-service", BuildOperationListenerService::class.java) {
                 this.parameters.getNegativeAvoidanceThreshold().set(extension.negativeAvoidanceThreshold)
             }