+++ doctor-plugin/src/main/java/com/osacky/doctor/internal/CoCaHelpers.kt
@@ -1,6 +1,19 @@
 package com.osacky.doctor.internal
 
-import org.gradle.api.invocation.Gradle
+import org.gradle.api.provider.ProviderFactory
 import org.gradle.util.GradleVersion
+import java.util.Optional
 
-fun Gradle.shouldUseCoCaClasses(): Boolean = GradleVersion.version(gradleVersion) >= GradleVersion.version("6.5")
+fun shouldUseCoCaClasses(): Boolean = isGradle65OrNewer()
+
+fun isGradle65OrNewer(): Boolean {
+    return GradleVersion.current() >= GradleVersion.version("6.5")
+}
+
+fun sysProperty(name: String, providers: ProviderFactory): Optional<String> {
+    if (isGradle65OrNewer()) {
+        val property = providers.systemProperty(name).forUseAtConfigurationTime()
+        return Optional.ofNullable(property.orNull)
+    }
+    return Optional.ofNullable(System.getProperty(name))
+}